refactor(domain): refactor directory structure
