+++ project/distribution/src/domain/repo/mod.rs
@@ -1,0 +1,5 @@
+mod model;
+mod repository;
+
+pub use model::*;
+pub use repository::*;
