+++ project/distribution/src/domain/user/mod.rs
@@ -1,0 +1,5 @@
+mod model;
+mod repository;
+
+pub use model::*;
+pub use repository::*;
