+++ project/distribution/src/service/repo.rs
@@ -38,3 +38,4 @@ pub async fn change_visibility(
         ),
     }
 }
+
